template_web/requirements.txt
# FastAPI y servidor ASGI
fastapi==0.115.12
uvicorn[standard]==0.34.2

# Templates
jinja2==3.1.6

# Base de datos (PostgreSQL async)
sqlalchemy==2.0.41
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Validación y serialización
pydantic==2.11.7
pydantic-settings==2.7.1

# Autenticación y seguridad
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-multipart==0.0.20

# HTTP client
httpx==0.28.1

# Forms
python-multipart==0.0.20

# Utilidades
python-dotenv==1.0.1

# Librería interna compartida (ajustar path según tu proyecto)
# -e ../repositorio
